warning fixed
boundedness check for transfer objects

renamings

renamings, part 2

renamings, part 3

estimator for TO capacity fixed

synchronization points in capacity estimation added

useless code removed

data structures for new genome model, part 2

data structures for new genome model, part 3

data structures for new genome model, part 4

serialization of new genome model

further fix for the capacity estimator of the particle array

renamings

+ transfer objects for new genome model
+ generate genomes in converter services

conversion description <-> TO for new genome model

bug fixes for description <-> TO conversion

+ GPU data structure for new genome model
+ data transfer TO -> GPU

wrong data type fixed

renamings

+ condensing genome data
+ GPU -> TO transfer

GPU -> TO transfer completed

adapt size estimator for new genome model

extend garbage collector for new genome model

+++ source/EngineGpuKernels/Base.cuh
@@ -42,7 +42,7 @@ struct PartitionData
     __inline__ __device__ int numElements() const { return endIndex - startIndex + 1; }
 };
 
-__device__ __inline__ PartitionData calcPartition(int numEntities, int index, int numIndices)
+__device__ __inline__ PartitionData calcPartition(uint64_t numEntities, uint64_t index, uint64_t numIndices)
 {
     PartitionData result;
     int entitiesByDivisions = numEntities / numIndices;
@@ -56,13 +56,13 @@ __device__ __inline__ PartitionData calcPartition(int numEntities, int index, in
     return result;
 }
 
-__device__ __inline__ PartitionData calcAllThreadsPartition(int numEntities)
+__device__ __inline__ PartitionData calcAllThreadsPartition(uint64_t numEntities)
 {
     return calcPartition(
         numEntities, threadIdx.x + blockIdx.x * blockDim.x, blockDim.x * gridDim.x);
 }
 
-__device__ __inline__ PartitionData calcBlockPartition(int numEntities)
+__device__ __inline__ PartitionData calcBlockPartition(uint64_t numEntities)
 {
     return calcPartition(numEntities, blockIdx.x, gridDim.x);
 }
